+++ crates/ruma-events/src/room/message/image.rs
@@ -2,7 +2,10 @@ use ruma_common::OwnedMxcUri;
 use serde::{Deserialize, Serialize};
 
 use super::FormattedBody;
-use crate::room::{EncryptedFile, ImageInfo, MediaSource};
+use crate::room::{
+    message::media_caption::{caption, formatted_caption},
+    EncryptedFile, ImageInfo, MediaSource,
+};
 
 /// The payload for an image message.
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -58,5 +61,20 @@ impl ImageMessageEventContent {
     /// as a shorthand for that, because it is very common to set this field.
     pub fn info(self, info: impl Into<Option<Box<ImageInfo>>>) -> Self {
         Self { info: info.into(), ..self }
+    }
+
+    /// Returns the caption for the image as defined by the [spec](https://spec.matrix.org/latest/client-server-api/#media-captions).
+    ///
+    /// In short, this is the `body` field if the `filename` field exists and has a different value,
+    /// otherwise the media file does not have a caption.
+    pub fn caption(&self) -> Option<&str> {
+        caption(&self.body, self.filename.as_deref())
+    }
+
+    /// Returns the formatted caption for the image as defined by the [spec](https://spec.matrix.org/latest/client-server-api/#media-captions).
+    ///
+    /// This is the same as `caption`, but returns the formatted body instead of the plain body.
+    pub fn formatted_caption(&self) -> Option<&FormattedBody> {
+        formatted_caption(&self.body, self.formatted.as_ref(), self.filename.as_deref())
     }
 }